+++ ChangeLog.md
@@ -7,6 +7,7 @@ https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 # 1.0.2 (dd-mm-yyyy)
   * Added localization for French, thanks to Nathalie Vedovotto for the contribution!
   * Fixed a bug that broke any download link
+  * Fixed a bug that caused an error when administrator tries to edit project settings
 
 # 1.0.1 (21-09-2021)
   * Added possibility to hide the top navigation bar to visitor users through the URL parameter "hideNav"